adicionando mudanças do modelo rest

+++ arquivos/resources/item.py
@@ -1,22 +1,26 @@
-import uuid
 from db import db
-from flask import request
 from models import ItemModel
 from schemas import ItemSchema
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required,get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 from flask_smorest import Blueprint, abort
 
 blp = Blueprint("Items",__name__,description="Operations on stores")
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
     @blp.response(200,ItemSchema)
     def get(self,item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self,item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401,message="Admin privilege required")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -43,6 +47,7 @@ class Item(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201,ItemSchema)
     def post(self,item_data):
